+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/VoiceChatRoomSetupActivity.kt
@@ -17,6 +17,9 @@ class VoiceChatRoomSetupActivity : AppCompatActivity() {
     }
 
     private val audioCodecOptions = listOf("OPUS", "PCMU")
+    private val audioBitRateOptions = listOf("UNDEFINED", "8", "16", "24", "32",
+            "64", "96", "128", "256")
+
     private val streamTypeOptions = listOf("BIDIRECTIONAL", "SINGLE-UP", "SINGLE-DOWN", "MULTI-DOWN")
     private val sdpSemanticsOptions = listOf("Unified Plan", "Plan B")
 
@@ -28,6 +31,8 @@ class VoiceChatRoomSetupActivity : AppCompatActivity() {
         start.setOnClickListener { startVoiceChat() }
         audioCodecSelection.name.text = "AUDIO CODEC"
         audioCodecSelection.spinner.setItems(audioCodecOptions)
+        audioBitRateSelection.name.text = "AUDIO BIT RATE"
+        audioBitRateSelection.spinner.setItems(audioBitRateOptions)
         streamTypeSelection.name.text = "STREAM TYPE"
         streamTypeSelection.spinner.setItems(streamTypeOptions)
         sdpSemanticsSelection.name.text = "SDP SEMANTICS"
@@ -43,12 +48,14 @@ class VoiceChatRoomSetupActivity : AppCompatActivity() {
 
         val streamType = selectedItem(streamTypeSelection.spinner)
         val audioCodec = selectedItem(audioCodecSelection.spinner)
+        val audioBitRate = selectedItem(audioBitRateSelection.spinner)
         val sdpSemantics = selectedItem(sdpSemanticsSelection.spinner)
 
         val intent = Intent(this, VoiceChatRoomActivity::class.java)
         intent.putExtra("CHANNEL_NAME", channelName)
         intent.putExtra("STREAM_TYPE", streamType)
         intent.putExtra("AUDIO_CODEC", audioCodec)
+        intent.putExtra("AUDIO_BIT_RATE", audioBitRate)
         intent.putExtra("SDP_SEMANTICS", sdpSemantics)
 
         startActivity(intent)